Add basic zsh configuration
This is antigen-based and largely taken from antigen's readme:
https://github.com/zsh-users/antigen/blob/64de2dcd95d6a8e879cd2244c763d99f0144e78e/README.mkd.
In addition to the basic items included there, I've also added a default
text editor and some Java-specific configuration for asdf. The Java
configuration may not actually work without the plugin first being
installed, so it will be interesting to see what happens there.

The readme has also been expanded with a few more details to get people
up and running.