Fix installation instructions
Previously I was hoping that `stow *` would be smart enough to run
through all directories and ignore files, but it looks like I need to be
a bit more explicit. This will likely wind up being a temporary approach
since I will want to be able to support multiple directories in the
future and adding another command for each directory isn't very
scalable.