Add apt, brew, and stow to installation script
This is primarily intended to automate the installation of `stow` which
is already used by our automation script (but does not necessarily come
pre-installed in all environments). In order to do that, we also need to
detect which package manager to use to install it. For now, brew and apt
are probably sufficient.

We've also updated the final status message to be more consistent with
the new messages.