Simplify git configuration when in codespaces
Codespaces already configures commit signing and user information, so we
can remove these values when we detect that we're in a Codespace. This
also means we don't have to worry about copying GPG keys around since
GitHub handles it for us.

Note: this won't work in zsh because of the `==` operator, which zsh
interprets differently than sh and bash.