Move to the root git directory during installation
Previously, the install script would only work if you invoked it from
the root of the git repository (since we have a couple commands that
are hard-coded for that, like our `stow` command). Now, we automatically
`cd` to the directory that contains `install.sh` (which simply does
nothing if we're already there).